codacy code style
