+++ database/alinosql/alinosql.go
@@ -1,41 +1,22 @@
 package alinosql
 
+//Listtables list tables
 func (alinosql *Alinosql) Listtables(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Deletetables delete tables
 func (alinosql *Alinosql) Deletetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Describetables Describe tables
 func (alinosql *Alinosql) Describetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Createtables create tables
 func (alinosql *Alinosql) Createtables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
-
-/*
-func (rackspaceserverless *Rackspaceserverless) Createfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Callfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Listfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-
-func (rackspaceserverless *Rackspaceserverless) Deletefunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-*/
