+++ database/alinosql/models.go
@@ -1,4 +1,5 @@
 package alinosql
 
+//Alinosql struct reperesnts  ali nosql service.
 type Alinosql struct {
 }
