+++ database/azurenosql/azurenosql.go
@@ -1,41 +1,22 @@
 package azurenosql
 
+//Listtables list tables
 func (azurenosql *Azurenosql) Listtables(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Deletetables delete tables
 func (azurenosql *Azurenosql) Deletetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Describetables Describe tables
 func (azurenosql *Azurenosql) Describetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Createtables create tables
 func (azurenosql *Azurenosql) Createtables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
-
-/*
-func (rackspaceserverless *Rackspaceserverless) Createfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Callfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Listfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-
-func (rackspaceserverless *Rackspaceserverless) Deletefunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-*/
