+++ database/azurenosql/models.go
@@ -1,4 +1,5 @@
 package azurenosql
 
+//Azurenosql struct reperesnts  Azure nosql service.
 type Azurenosql struct {
 }
