+++ database/digioceannosql/digioceannosql.go
@@ -1,41 +1,22 @@
 package digioceannosql
 
+//Listtables list tables
 func (digioceannosql *Digioceannosql) Listtables(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Deletetables delete tables
 func (digioceannosql *Digioceannosql) Deletetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Describetables Describe tables
 func (digioceannosql *Digioceannosql) Describetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Createtables create tables
 func (digioceannosql *Digioceannosql) Createtables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
-
-/*
-func (rackspaceserverless *Rackspaceserverless) Createfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Callfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Listfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-
-func (rackspaceserverless *Rackspaceserverless) Deletefunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-*/
