+++ database/digioceannosql/models.go
@@ -1,4 +1,5 @@
 package digioceannosql
 
+//Digioceannosql struct reperesnts  Digiocean nosql service.
 type Digioceannosql struct {
 }
