+++ database/openstacknosql/models.go
@@ -1,4 +1,5 @@
 package openstacknosql
 
+//Openstacknosql struct reperesnts  openatack nosql service.
 type Openstacknosql struct {
 }
