+++ database/openstacknosql/openstacknosql.go
@@ -1,41 +1,22 @@
 package openstacknosql
 
+//Listtables list tables
 func (openstacknosql *Openstacknosql) Listtables(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Deletetables delete tables
 func (openstacknosql *Openstacknosql) Deletetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Describetables Describe tables
 func (openstacknosql *Openstacknosql) Describetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Createtables create tables
 func (openstacknosql *Openstacknosql) Createtables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
-
-/*
-func (rackspaceserverless *Rackspaceserverless) Createfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Callfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Listfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-
-func (rackspaceserverless *Rackspaceserverless) Deletefunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-*/
