+++ database/rackspacenosql/models.go
@@ -1,4 +1,5 @@
 package rackspacenosql
 
+//Rackspacenosql struct reperesnts  Rackspace nosql service.
 type Rackspacenosql struct {
 }
