+++ database/rackspacenosql/rackspacenosql.go
@@ -1,41 +1,22 @@
 package rackspacenosql
 
+//Listtables list tables
 func (rackspacenosql *Rackspacenosql) Listtables(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Deletetables delete tables
 func (rackspacenosql *Rackspacenosql) Deletetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Describetables Describe tables
 func (rackspacenosql *Rackspacenosql) Describetables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
 
+//Createtables create tables
 func (rackspacenosql *Rackspacenosql) Createtables(request interface{}) (resp interface{}, err error) {
 	return resp, err
 }
-
-/*
-func (rackspaceserverless *Rackspaceserverless) Createfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Callfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (rackspaceserverless *Rackspaceserverless) Listfunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-
-func (rackspaceserverless *Rackspaceserverless) Deletefunction(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-*/
