+++ machinelearning/alimachinelearning/alimachinelearning.go
@@ -1,20 +1,24 @@
 package alimachinelearning
 
+//CreateMLModel creates model.
 func (alimachinelearning *Alimachinelearning) CreateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//DeleteMLModel delete model.
 func (alimachinelearning *Alimachinelearning) DeleteMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//GetMLModel get model.
 func (alimachinelearning *Alimachinelearning) GetMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//UpdateMLModel  Update model.
 func (alimachinelearning *Alimachinelearning) UpdateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
