+++ machinelearning/alimachinelearning/models.go
@@ -1,4 +1,5 @@
 package alimachinelearning
 
+//Alimachinelearning struct reperesnts ali machine learning service.
 type Alimachinelearning struct {
 }
