+++ machinelearning/azuremachinelearning/azuremachinelearning.go
@@ -1,20 +1,24 @@
 package azuremachinelearning
 
+//CreateMLModel creates model.
 func (azuremachinelearning *Azuremachinelearning) CreateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//DeleteMLModel delete model.
 func (azuremachinelearning *Azuremachinelearning) DeleteMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//GetMLModel get model.
 func (azuremachinelearning *Azuremachinelearning) GetMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//UpdateMLModel  Update model.
 func (azuremachinelearning *Azuremachinelearning) UpdateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
