+++ machinelearning/azuremachinelearning/models.go
@@ -1,4 +1,5 @@
 package azuremachinelearning
 
+//Azuremachinelearning struct reperesnts azure machine learning service.
 type Azuremachinelearning struct {
 }
