+++ machinelearning/digioceanmachinelearning/awsmachinelearning.go
@@ -1,20 +1,24 @@
-package awsmachinelearning
+package digioceanmachinelearning
 
-func (awsmachinelearning *Awsmachinelearning) CreateMLModel(request interface{}) (resp interface{}, err error) {
-
-	return resp, err
-}
-
-func (awsmachinelearning *Awsmachinelearning) DeleteMLModel(request interface{}) (resp interface{}, err error) {
+//CreateMLModel creates model.
+func (digioceanmachinelearning *Digioceanmachinelearning) CreateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
-func (awsmachinelearning *Awsmachinelearning) GetMLModel(request interface{}) (resp interface{}, err error) {
+//DeleteMLModel delete model.
+func (digioceanmachinelearning *Digioceanmachinelearning) DeleteMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//GetMLModel get model.
+func (digioceanmachinelearning *Digioceanmachinelearning) GetMLModel(request interface{}) (resp interface{}, err error) {
+
+	return resp, err
+}
+
+//UpdateMLModel  Update model.
 func (awsmachinelearning *Awsmachinelearning) UpdateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
