+++ machinelearning/googlemachinelearning/googlemachinelearning.go
@@ -1,20 +1,25 @@
 package googlemachinelearning
 
+
+//CreateMLModel creates model.
 func (googlemachinelearning *Googlemachinelearning) CreateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//DeleteMLModel delete model.
 func (googlemachinelearning *Googlemachinelearning) DeleteMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//GetMLModel get model.
 func (googlemachinelearning *Googlemachinelearning) GetMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//UpdateMLModel  Update model.
 func (googlemachinelearning *Googlemachinelearning) UpdateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
