+++ machinelearning/googlemachinelearning/models.go
@@ -1,4 +1,5 @@
 package googlemachinelearning
 
+//Googlemachinelearning struct reperesnts google machine learning service.
 type Googlemachinelearning struct {
 }
