+++ machinelearning/openstackmachinelearning/models.go
@@ -1,4 +1,5 @@
 package openstackmachinelearning
 
+//Openstackmachinelearning struct reperesnts openstack machine learning service.
 type Openstackmachinelearning struct {
 }
