+++ machinelearning/openstackmachinelearning/openstackmachinelearning.go
@@ -1,20 +1,25 @@
 package openstackmachinelearning
 
+
+//CreateMLModel creates model.
 func (openstackmachinelearning *Openstackmachinelearning) CreateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//DeleteMLModel delete model.
 func (openstackmachinelearning *Openstackmachinelearning) DeleteMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//GetMLModel get model.
 func (openstackmachinelearning *Openstackmachinelearning) GetMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//UpdateMLModel  Update model.
 func (openstackmachinelearning *Openstackmachinelearning) UpdateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
