+++ machinelearning/rackspacemachinelearning/models.go
@@ -1,4 +1,5 @@
 package rackspacemachinelearning
 
+//Rackspacemachinelearning struct reperesnts rackspace machine learning service.
 type Rackspacemachinelearning struct {
 }
