+++ machinelearning/rackspacemachinelearning/rackspacemachinelearning.go
@@ -1,20 +1,24 @@
 package rackspacemachinelearning
 
+//CreateMLModel creates model.
 func (rackspacemachinelearning *Rackspacemachinelearning) CreateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//DeleteMLModel delete model.
 func (rackspacemachinelearning *Rackspacemachinelearning) DeleteMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//GetMLModel get model.
 func (rackspacemachinelearning *Rackspacemachinelearning) GetMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//UpdateMLModel  Update model.
 func (rackspacemachinelearning *Rackspacemachinelearning) UpdateMLModel(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
