+++ serverless/aliserverless/aliserverless.go
@@ -1,25 +1,30 @@
 package aliserverless
 
+//Getfunction  describe serverless function.
 func (aliserverless *Aliserverless) Getfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Createfunction  create serverless function.
 func (aliserverless *Aliserverless) Createfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Callfunction  call serverless function.
 func (aliserverless *Aliserverless) Callfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Listfunction  list serverless function.
 func (aliserverless *Aliserverless) Listfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Deletefunction  delete serverless function.
 func (aliserverless *Aliserverless) Deletefunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
