+++ serverless/aliserverless/models.go
@@ -1,4 +1,5 @@
 package aliserverless
 
+//Aliserverless struct reperesnts ali  serverless service.
 type Aliserverless struct {
 }
