+++ serverless/azureserverless/models.go
@@ -1,4 +1,5 @@
 package azureserverless
 
+//Azureserverless struct reperesnts azure  serverless service.
 type Azureserverless struct {
 }
