+++ serverless/digioceanserverless/models.go
@@ -1,4 +1,5 @@
 package digioceanserverless
 
+//Digioceanserverless struct reperesnts Digiocean serverless service.
 type Digioceanserverless struct {
 }
