+++ serverless/lambda/models.go
@@ -1,5 +1,6 @@
 package lambda
 
+//Lambda struct reperesnts aws serverless service.
 type Lambda struct {
 }
 
@@ -8,6 +9,7 @@ type Deletefunction struct {
 	Qualifier    string
 }
 
+//Createfunction struct represents aws serverless Create function.
 type Createfunction struct {
 	code Code `json:"Code"`
 
@@ -27,31 +29,38 @@ type Createfunction struct {
 	vpcConfig     VpcConfig     `json:"VpcConfig"`
 }
 
+//Tags struct represents Createfunction parameters.
 type Tags struct {
 	String string `json:"string"`
 }
 
+//TracingConfig struct represents Createfunction parameters.
 type TracingConfig struct {
 	Mode string `json:"Mode"`
 }
 
+//VpcConfig struct represents Createfunction parameters.
 type VpcConfig struct {
 	SecurityGroupIds []string `json:"SecurityGroupIds"`
 	SubnetIds        []string `json:"SubnetIds"`
 }
 
+//Variables struct represents Createfunction parameters.
 type Variables struct {
 	String string `json:"string"`
 }
 
+//Environment struct represents Createfunction parameters.
 type Environment struct {
 	variables Variables `json:"Variables"`
 }
 
+//DeadLetterConfig struct represents Createfunction parameters.
 type DeadLetterConfig struct {
 	TargetArn string `json:"TargetArn"`
 }
 
+//Code struct represents Createfunction parameters.
 type Code struct {
 	S3Bucket        string `json:"S3Bucket"`
 	S3Key           string `json:"S3Key"`
