+++ serverless/openstackserverless/models.go
@@ -1,4 +1,5 @@
 package openstackserverless
 
+//Openstackserverless struct reperesnts openstack  serverless service.
 type Openstackserverless struct {
 }
