+++ serverless/openstackserverless/openstackserverless.go
@@ -1,25 +1,31 @@
 package openstackserverless
 
+
+//Getfunction  describe serverless function.
 func (openstackserverless *Openstackserverless) Getfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Createfunction  create serverless function.
 func (openstackserverless *Openstackserverless) Createfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Callfunction  call serverless function.
 func (openstackserverless *Openstackserverless) Callfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Listfunction  list serverless function.
 func (openstackserverless *Openstackserverless) Listfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Deletefunction  delete serverless function.
 func (openstackserverless *Openstackserverless) Deletefunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
