+++ serverless/rackspaceserverless/models.go
@@ -1,4 +1,5 @@
 package rackspaceserverless
 
+//Rackspaceserverless struct reperesnts Rackspace  serverless service.
 type Rackspaceserverless struct {
 }
