+++ serverless/rackspaceserverless/rackspaceserverless.go
@@ -1,25 +1,31 @@
 package rackspaceserverless
 
+
+//Getfunction  describe serverless function.
 func (rackspaceserverless *Rackspaceserverless) Getfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Createfunction  create serverless function.
 func (rackspaceserverless *Rackspaceserverless) Createfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Callfunction  call serverless function.
 func (rackspaceserverless *Rackspaceserverless) Callfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Listfunction  list serverless function.
 func (rackspaceserverless *Rackspaceserverless) Listfunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
 }
 
+//Deletefunction  delete serverless function.
 func (rackspaceserverless *Rackspaceserverless) Deletefunction(request interface{}) (resp interface{}, err error) {
 
 	return resp, err
